requirements.txt
numpy==1.21.2
opencv_python==4.5.3.56
paddlepaddle==2.1.3
paddleseg==2.3.0
PyQt5==5.15.6
SimpleITK==2.1.1
